requirements.txt
streamlit==1.39.0
pandas==2.2.2
pillow==10.3.0

google-genai==0.3.0
anthropic==0.30.1
openai==1.35.7

notion-client==2.2.1

reportlab==4.1.0
qrcode==7.4.2

requests==2.32.3
